Separate extension, server and worker paths
