+++ src/common/completion_docs.ts
