+++ src/extension/browser/extension_browser.ts
@@ -30,7 +30,7 @@ import {
 } from 'vscode-languageclient/browser';
 
 // import { WorkerConnection } from "../../worker/browser/worker_connection";
-import {WorkerConnection} from '../../worker/browser/workerless_worker';
+import {WorkerConnection} from './workerless_worker';
 import {setupSubscriptions} from '../subscriptions';
 import {MalloyConfig} from '../types';
 import {connectionManager} from './connection_manager';
