+++ src/extension/node/commands/edit_connections.ts
@@ -30,7 +30,7 @@ import {
   ConnectionPanelMessage,
   ConnectionServiceAccountKeyRequestStatus,
   ConnectionTestStatus,
-} from '../../message_types';
+} from '../../../common/message_types';
 import {WebviewMessageManager} from '../../webview_message_manager';
 import {connectionManager} from '../connection_manager';
 import {
