+++ src/extension/node/extension_node.ts
@@ -32,13 +32,13 @@ import {
 } from 'vscode-languageclient/node';
 import {editConnectionsCommand} from './commands/edit_connections';
 import {ConnectionsProvider} from '../tree_views/connections_view';
-import {WorkerConnection} from '../../worker/node/worker_connection';
+import {WorkerConnection} from './worker_connection';
 import {MalloyConfig} from '../types';
 import {connectionManager} from './connection_manager';
 import {setupSubscriptions} from '../subscriptions';
 import {initFileMessaging, VSCodeURLReader} from '../utils';
 import {MALLOY_EXTENSION_STATE} from '../state';
-import {BaseWorker} from '../../worker/types';
+import {BaseWorker} from '../../common/worker_message_types';
 
 let client: LanguageClient;
 let worker: BaseWorker;
