+++ src/extension/state.ts
@@ -23,7 +23,7 @@
 
 import vscode, {TextDocument, WebviewPanel} from 'vscode';
 import {Result} from '@malloydata/malloy';
-import {QueryPanelMessage} from './message_types';
+import {QueryPanelMessage} from '../common/message_types';
 import {WebviewMessageManager} from './webview_message_manager';
 
 export interface RunState {
