+++ src/extension/webviews/help_page/entry.ts
@@ -27,7 +27,7 @@ import ReactDOM from 'react-dom';
 import React from 'react';
 import {App} from './App';
 import {getVSCodeAPI, HelpVSCodeContext} from './help_vscode_context';
-import {HelpPanelMessage} from '../../message_types';
+import {HelpPanelMessage} from '../../../common/message_types';
 
 (() => {
   const vscode = getVSCodeAPI<void, HelpPanelMessage>();
