+++ src/extension/webviews/help_view.ts
@@ -24,7 +24,7 @@
 import * as vscode from 'vscode';
 import {Utils} from 'vscode-uri';
 import {getWebviewHtml} from '../webviews';
-import {HelpMessageType, HelpPanelMessage} from '../message_types';
+import {HelpMessageType, HelpPanelMessage} from '../../common/message_types';
 import {WebviewMessageManager} from '../webview_message_manager';
 
 export class HelpViewProvider implements vscode.WebviewViewProvider {
