+++ src/extension/webviews/query_page/DownloadButton.tsx
@@ -23,7 +23,7 @@
 
 import React, {useState} from 'react';
 import styled from 'styled-components';
-import {QueryDownloadOptions} from '../../message_types';
+import {QueryDownloadOptions} from '../../../common/message_types';
 import {Popover} from '../components/Popover';
 import {DownloadForm} from './DownloadForm';
 import DownloadIcon from '../assets/download_hover.svg';
