+++ src/extension/webviews/query_page/query_vscode_context.ts
@@ -21,7 +21,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-import {QueryPanelMessage} from '../../message_types';
+import {QueryPanelMessage} from '../../../common/message_types';
 import {makeUseVSCodeContext, makeVSCodeContext} from '../vscode_context';
 
 export const QueryVSCodeContext = makeVSCodeContext<void, QueryPanelMessage>();
