+++ src/server/hover/hover.ts
@@ -24,7 +24,7 @@
 import {Hover, HoverParams, MarkupKind} from 'vscode-languageserver/node';
 import {TextDocument} from 'vscode-languageserver-textdocument';
 
-import {COMPLETION_DOCS} from '../completions/completion_docs';
+import {COMPLETION_DOCS} from '../../common/completion_docs';
 import {parseWithCache} from '../parse_cache';
 
 export const getHover = (
