+++ src/worker/node/message_handler.ts
@@ -24,7 +24,11 @@
 import {log} from '../logger';
 import {cancelQuery, runQuery} from '../run_query';
 import {downloadQuery} from './download_query';
-import {Message, MessageHandler, WorkerMessage} from '../types';
+import {
+  Message,
+  MessageHandler,
+  WorkerMessage,
+} from '../../common/worker_message_types';
 import {refreshConfig} from '../refresh_config';
 import {ConnectionManager} from '../../common/connection_manager';
 import {WorkerURLReader, fetchCellData} from './files';
